MR-103: create Form
