+++ src/Components/Members/TableMember/index.js
@@ -1,7 +1,14 @@
 import React from 'react';
 import styles from './table-member.module.css';
+import { useHistory } from 'react-router-dom/cjs/react-router-dom.min';
 
 const TableMember = ({ members }) => {
+  const history = useHistory();
+
+  const handleClick = (id) => {
+    history.push(`members/membersForm/${id}`);
+  };
+
   return (
     <div className={styles.container}>
       <table className={styles.table}>
@@ -29,7 +36,7 @@ const TableMember = ({ members }) => {
               <td>{member.postalCode}</td>
               <td>{member.membership}</td>
               <td className={styles.rowActions}>
-                <button>Edit</button>
+                <button onClick={() => handleClick(member._id)}>Edit</button>
                 <button>Delete</button>
               </td>
             </tr>
